Updating CONTRIBUTING.md; splitting test_remote_configuration.py into provider and requirer part
